requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
Deprecated==1.2.3
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
PyGithub==1.43.2
Pygments==2.2.0
PyJWT==1.6.4
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
